requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
daphne==2.5.0
Django==4.0.6
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
outcome==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.1
requests==2.28.1
selenium==4.3.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.10
wsproto==1.1.0
zope.interface==5.4.0
